requirements.txt
django-model-utils>=2.0

# PROTECTED REGION ID(ngosubs.requirements.base.user) ENABLED START
# Additional requirements go here
django-ngoshop
# PROTECTED REGION END
